+++ src/main/java/com/inq/wishhair/wesharewishhair/global/mail/EmailSender.java
@@ -1,4 +1,4 @@
-package com.inq.wishhair.wesharewishhair.global.mail.utils;
+package com.inq.wishhair.wesharewishhair.global.mail;
 
 import java.io.UnsupportedEncodingException;
